# src/AppForSEII2526.Web/OpenAPIs/swagger1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AppForMovies.API",
    "description": "This API provides services for renting and purchasing movies",
    "contact": {
      "name": "Software Engineering II Team",
      "email": "isii@on.uclm.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "v1"
  },
  "paths": {
    "/api/Alquileres/DetalleAlquiler": {
      "get": {
        "tags": [
          "Alquileres"
        ],
        "operationId": "GetDetalleHerramientasParaAlquiler",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AlquileresParaDetalleDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlquileresParaDetalleDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlquileresParaDetalleDTO"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Alquileres/CrearAlquiler": {
      "post": {
        "tags": [
          "Alquileres"
        ],
        "operationId": "CreateAlquiler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearAlquilerDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearAlquilerDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CrearAlquilerDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AlquileresParaDetalleDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlquileresParaDetalleDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlquileresParaDetalleDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AlquilarItemsDTO": {
        "type": "object",
        "properties": {
          "nombreItem": {
            "type": "string",
            "nullable": true
          },
          "materialItem": {
            "type": "string",
            "nullable": true
          },
          "precioItem": {
            "type": "number",
            "format": "float"
          },
          "cantidadItem": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "AlquileresParaDetalleDTO": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "apellidos": {
            "type": "string",
            "nullable": true
          },
          "direccion": {
            "type": "string",
            "nullable": true
          },
          "fechaAlquiler": {
            "type": "string",
            "format": "date"
          },
          "precioTotal": {
            "type": "number",
            "format": "float"
          },
          "fechaInicio": {
            "type": "string",
            "format": "date"
          },
          "fechaFinal": {
            "type": "string",
            "format": "date"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlquilarItemsDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CrearAlquilerDTO": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "apellidos": {
            "type": "string",
            "nullable": true
          },
          "metodoPagoId": {
            "type": "integer",
            "format": "int32"
          },
          "direccion": {
            "type": "string",
            "nullable": true
          },
          "telefono": {
            "type": "string",
            "nullable": true
          },
          "correo": {
            "type": "string",
            "nullable": true
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CrearAlquilerItemDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CrearAlquilerItemDTO": {
        "type": "object",
        "properties": {
          "herramientaId": {
            "type": "integer",
            "format": "int32"
          },
          "herramientaCantidad": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    }
  }
}